requirement.txt
pandas>=1.3.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
pyodbc>=4.0.30
shareplum>=0.5.1